requirements.txt
jpype1
pywin32